+++ src/co/insou/byteme/instruct/ByteMe2BInstruction.java
@@ -3,7 +3,7 @@ package co.insou.byteme.instruct;
 public interface ByteMe2BInstruction extends ByteMeInstruction {
 
     @Override
-    default byte parameters() {
+    default int parameters() {
         return 1;
     }
 }
