+++ src/co/insou/byteme/instruct/def/Decr_25.java
@@ -7,7 +7,7 @@ public class Decr_25 implements ByteMeInstruction {
 
     @Override
     public void execute(ByteMeRuntime runtime) {
-        runtime.stack().push((byte) (runtime.stack().pop() - 1));
+        runtime.stack().push(runtime.stack().pop() - 1);
     }
 
 }
