+++ src/co/insou/byteme/instruct/def/GotoSIf_31.java
@@ -8,7 +8,7 @@ public class GotoSIf_31 implements ByteMeInstruction {
     @Override
     public void execute(ByteMeRuntime runtime) {
         if (runtime.stack().pop() != 0) {
-            byte index = runtime.stack().pop();
+            int index = runtime.stack().pop();
             runtime.setPC(index);
         }
     }
