+++ src/co/insou/byteme/vm/ByteMeProgram.java
@@ -2,7 +2,7 @@ package co.insou.byteme.vm;
 
 public interface ByteMeProgram {
 
-    byte[] code();
+    int[] code();
 
     int len();
